// Frontend/Admin/src/app/models/product.ts
import { FileHandle } from "./file-handle";

export class Product {
    product_name: string;
    category_id: number;
    avatar: string;
    img1: string;
    img2: string;
    img3: string;
    description: string;
    price: number;
    discount: number;
    quantity: number;
    created_by: string;
    updated_by: string;

}